feat: preprocessing, allow short dates on wallet
Let wallet csv to shorten the date for each row possibly till only date
is given. In which case the converter will use the year and the month
info defined on the first row.

Define `preprocess` function provided by converters to consolidate this
date setting and the preprocessing from revolut converter.